Move LinalgOpInfo to Codegen/Common
