+++ compiler/src/iree/compiler/Codegen/Common/LinalgOpInfo.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Codegen/LLVMCPU/LinalgOpInfo.h"
+#include "iree/compiler/Codegen/Common/LinalgOpInfo.h"
 
 #include "mlir/Dialect/Linalg/IR/LinalgInterfaces.h"
